requirements.txt
pytest~=8.3.2
pycodestyle
PyYAML~=6.0.2